Сделал, что все работает, в том числе админка. Убрал лишние файлы. Исправил код, где нужно. СОбрал из кусков нужное.
