+++ bot/.env.example
@@ -1,1 +1,2 @@
-BOT_CONFIG__APP__TOKEN=tokentelegi
+BOT_CONFIG__APP__TOKEN="token"
+BOT_CONFIG__API__BASE_URL="http://127.0.0.1"
