+++ bot/bot.py
@@ -3,38 +3,29 @@ import logging
 from aiogram import Bot, Dispatcher
 from aiogram.fsm.storage.memory import MemoryStorage
 
-# from routers import tree_commands
 from core.config import settings
-from routers import main_menu, navigation
-from handlers import main_button
-
+from routers import admin, main_menu, navigation, tree_commands
 
 
 async def main():
-
     logging.basicConfig(
         level=logging.INFO,
         format="%(asctime)s - %(levelname)s - %(name)s - %(message)s",
     )
-    
+    if not settings.app.token:
+        raise ValueError(
+            "Не найден токен бота. Пожалуйста,"
+            "добавьте BOT_CONFIG__APP__TOKEN в .env файл."
+        )
     dp = Dispatcher(storage=MemoryStorage())
-
-    bot = Bot(token=settings.app.token)
-
-    # Andrey
-    # Создаем бота
-    #bot = Bot(
-    #    token=TOKEN,
-    #    session=AiohttpSession(),
-    #    default=DefaultBotProperties(parse_mode=ParseMode.HTML)
-    #)
-
+    bot = Bot(
+        token=settings.app.token,
+        
+    )
     dp.include_router(main_menu.router)
     dp.include_router(navigation.router)
-
-    # dp.include_router(main_button.router) # здесь подключаем хендлеры
-    # dp.include_router(tree_commands.router) # подключаем роутер с деревом
-
+    dp.include_router(tree_commands.router)
+    dp.include_router(admin.router)
     await dp.start_polling(bot)
