--- bot/handlers/test_hand.py
@@ -1,8 +1,0 @@
-from api import get_api_data
-
-
-
-
-CONTENT_URL = '/bot_menu/{button_id}/get-content'
-IMG_URL = '/bot_menu/{button_id}/get-image-file'
-CHILD_URL = '/bot_menu/{button_id}/get-child-buttons'
