+++ bot/utils/tree_utils.py
@@ -1,4 +1,4 @@
-from api.api_for_tree import get_api_tree
+from api.api_service import get_child_buttons
 
 
 def format_text(data, prefix="", is_last=True):
@@ -17,7 +17,7 @@ def format_text(data, prefix="", is_last=True):
     return result
 
 async def render_recursive(button_id):    
-    child_buttons = await get_api_tree(button_id)    
+    child_buttons = await get_child_buttons(button_id)    
     for button in child_buttons:
         button["children"] = await render_recursive(button["id"])
     return child_buttons
